+++ bin/nanocld/changelog.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `--store-addr` options from the command line to specify the store address
 - Endpoint `GET /metrics/{key}/inspect` to get details about a metric
 - Endpoint `GET /event/{key}/inspect` to get details about an event
+- Generic RepositoryCountBy interface
+- Endpoint `GET /metrics/count` to get the number of metrics
+- status field for VmSummary, CargoSummary
 
 ### Fixed